Refactor detailed exceptions block
